Rename UI tools file
