+++ src/filesystems/local/deployTools.ts
@@ -4,7 +4,7 @@ import vscode, { Uri, WorkspaceFolder } from 'vscode';
 import { instance } from '../../instantiate';
 import { LocalLanguageActions } from './LocalLanguageActions';
 import { Deployment } from './deployment';
-import { VscodeTools } from '../../ui/tools';
+import { VscodeTools } from '../../ui/vscodeTools';
 import { DeploymentMethod } from '../../api/types';
 import { DeploymentParameters } from '../../typings';
